Refactored monsterframes
I've refactored the monster frame so all icons are standard and will load consistently. This was done by using code from addon.ipf/mobinfobyskill/mobinfobyskill.lua as noted in the changed file.
I've added the movement type to the frame.
I've also added some conditional padding so that the info displays correctly when an elite monster is targeted. They have different frames, so the spacing works out odd.
I've left the text portion of the size commented in case people prefer that over the icons. I like the icons myself.
I've also left the exp part commented. Still can't get it to work right when an elite monster shares a name with a normal monster.